Added Elastic Search Client jar
