--- bin/application.properties
@@ -1,4 +1,0 @@
-security.user.name=admin
-security.user.password=password
-management.security.roles=SUPERUSER
-management.security.enabled=true
