+++ src/main/java/com/syntel/hackathon/log/aggregator/configuration/Application.java
@@ -1,0 +1,14 @@
+package com.syntel.hackathon.log.aggregator.configuration;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.ConfigurableApplicationContext;
+
+@SpringBootApplication(scanBasePackages = { "com.syntel.hackathon.log.aggregator" })
+public class Application {
+    public static void main(String[] args) {
+        ConfigurableApplicationContext context = SpringApplication.run(Application.class, args);
+        ApplicationContextProvider.setApplicationContext(context);
+    }
+
+}
